Refactoring ValueResolver so that conversion methods are in another class
